+++ js/src/whalesong/message.js
@@ -35,23 +35,11 @@ export class MessageManager extends ModelManager {
     });
   }
 
-  getSubmanager(name) {
-    try {
-      return super.getSubmanager(name);
-    } catch (err) {
-      if (name === 'msgInfo') {
-        try {
-          return manager.getSubmanager('messageInfos').getSubmanager(this.model.id._serialized);
-        } catch (err2) {}
-      }
-
-      throw err;
-    }
-  }
-
   @command
   async fetchInfo() {
-    return await manager.getSubmanager('messageInfos').fetchByMessage(this.model);
+    let result = await manager.getSubmanager('messageInfos').fetchByMessage(this.model);
+    this.addSubmanager('msgInfo', manager.getSubmanager('messageInfos').getSubmanager(this.model.id._serialized));
+    return result;
   }
 }