+++ js/src/whalesong/stream.js
@@ -6,27 +6,6 @@ import {
 } from '../manager.js';
 
 
-export const STATE = {
-  OPENING: "OPENING",
-  PAIRING: "PAIRING",
-  UNPAIRED: "UNPAIRED",
-  UNPAIRED_IDLE: "UNPAIRED_IDLE",
-  CONNECTED: "CONNECTED",
-  TIMEOUT: "TIMEOUT",
-  CONFLICT: "CONFLICT",
-  UNLAUNCHED: "UNLAUNCHED",
-  PROXYBLOCK: "PROXYBLOCK",
-  TOS_BLOCK: "TOS_BLOCK",
-  SMB_TOS_BLOCK: "SMB_TOS_BLOCK"
-}
-
-export const STREAM = {
-  DISCONNECTED: "DISCONNECTED",
-  SYNCING: "SYNCING",
-  RESUMING: "RESUMING",
-  CONNECTED: "CONNECTED"
-}
-
 export class StreamManager extends ModelManager {
 
   static mapModel(item) {
@@ -35,10 +14,9 @@ export class StreamManager extends ModelManager {
       canSend: item.canSend,
       hasSynced: item.hasSynced,
       isIncognito: item.isIncognito,
-      lastPhoneMessage: item.lastPhoneMessage,
+//      lastPhoneMessage: item.lastPhoneMessage,
       launchGeneration: item.launchGeneration,
       launched: item.launched,
-      pokeable: null,
       retryTimestamp: item.retryTimestamp,
       state: item.state,
       stream: item.stream,